Split JS/TSX React tests and language ids
Register javascriptreact (.jsx) and typescriptreact (.tsx) separately
from plain JS/TS, and move JSX coverage into dedicated test files.
