+++ test/javascript.test.ts
@@ -244,39 +244,3 @@ test("javascript: extracts generator and exported arrow bodies", ({
       └─ done()
   `);
 });
-
-test("javascript: tracks JSX components as nested calls", ({
-  expectCallstack,
-}) => {
-  expectCallstack(
-    `
-      export function App() {
-        setup();
-        return (
-    -     <Layout>
-    -       <Header />
-    -     </Layout>
-    +     <Shell>
-    +       <Header />
-    +       <Sidebar />
-    +     </Shell>
-        );
-      }
-      function setup() {}
-    - function Layout(props) { return null; }
-    + function Shell(props) { return null; }
-      function Header() { return null; }
-    + function Sidebar() { return null; }
-    `,
-    "App",
-    { file: "app.jsx" },
-  ).toEqual(`
-      App()
-      ├─ setup()
-    - ├─ Layout(props)
-    - │  └─ Header()
-    + └─ Shell(props)
-    +    ├─ Header()
-    +    └─ Sidebar()
-  `);
-});
